requirements.txt
arrow==0.14.7
ics==0.7
Pillow==7.0.0
psycopg2==2.8.4
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dateutil==2.8.1
six==1.14.0
TatSu==4.4.0
